+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.0.0] - [2018-07-01]
 ### Added 
 - Created and finalized the static website and the blog. Next is to setup different environment.
+- Added Judy and Gurdeep to the About and Press pages.
 
 ### Changed
 - Removed TD mention from our front page because we haven't heard back from TD.